fix: #520 (moving types out, fixing soundness) (#526)
* chore: move IO types

* chore: move everything out of types.d.ts, some compile errs

* fix: some typerrors, not all

* chore: fix all other type errors

* fix: declare returntype for insertPending

* chore: nimo requests

* refactor: reorganize types; rename type modules

* chore: prepare for merge
