+++ packages/browser-ui/src/App.tsx
@@ -11,7 +11,7 @@ import {
   stateInitial,
   stepUntilConvergence,
   showError,
-  PenroseError
+  PenroseError,
 } from "@penrose/core";
 
 /**
@@ -24,7 +24,7 @@ export const DownloadSVG = (
   title = "illustration"
 ): void => {
   const blob = new Blob([svg.outerHTML], {
-    type: "image/svg+xml;charset=utf-8"
+    type: "image/svg+xml;charset=utf-8",
   });
   const url = URL.createObjectURL(blob);
   const downloadLink = document.createElement("a");
@@ -64,7 +64,7 @@ class App extends React.Component<any, ICanvasState> {
     penroseVersion: "",
     showInspector: true,
     files: null,
-    connected: false
+    connected: false,
   };
   public readonly buttons = React.createRef<ButtonBar>();
   public readonly canvasRef = React.createRef<HTMLDivElement>();
@@ -76,23 +76,23 @@ class App extends React.Component<any, ICanvasState> {
   // same as onCanvasState but doesn't alter timeline or involve optimization
   // used only in modshapes
   public modCanvas = async (canvasState: PenroseState) => {
-    await new Promise(r => setTimeout(r, 1));
+    await new Promise((r) => setTimeout(r, 1));
 
     this.setState({
       data: canvasState,
-      processedInitial: true
+      processedInitial: true,
     });
     this.renderCanvas(canvasState);
   };
   public onCanvasState = async (canvasState: PenroseState) => {
     // HACK: this will enable the "animation" that we normally expect
-    await new Promise(r => setTimeout(r, 1));
+    await new Promise((r) => setTimeout(r, 1));
 
     this.setState({
       data: canvasState,
       history: [...this.state.history, canvasState],
       processedInitial: true,
-      error: null
+      error: null,
     });
     this.renderCanvas(canvasState);
     const { autostep } = this.state;
@@ -108,25 +108,31 @@ class App extends React.Component<any, ICanvasState> {
   };
   public downloadState = (): void => {
     const state = this.state.data;
-    const params = {
-      ...state.params,
-      energyGraph: {},
-      xsVars: [],
-      constrWeightNode: undefined,
-      epWeightNode: undefined,
-      graphs: undefined
-    };
-    const content = JSON.stringify({ ...state, params });
-    const blob = new Blob([content], {
-      type: "text/json"
-    });
-    const url = URL.createObjectURL(blob);
-    const downloadLink = document.createElement("a");
-    downloadLink.href = url;
-    downloadLink.download = `state.json`;
-    document.body.appendChild(downloadLink);
-    downloadLink.click();
-    document.body.removeChild(downloadLink);
+    if (state) {
+      const params = {
+        ...state.params,
+        energyGraph: {},
+        xsVars: [],
+        constrWeightNode: undefined,
+        epWeightNode: undefined,
+        graphs: undefined,
+      };
+      const content = JSON.stringify({ ...state, params });
+      const blob = new Blob([content], {
+        type: "text/json",
+      });
+      const url = URL.createObjectURL(blob);
+      const downloadLink = document.createElement("a");
+      downloadLink.href = url;
+      downloadLink.download = `state.json`;
+      document.body.appendChild(downloadLink);
+      downloadLink.click();
+      document.body.removeChild(downloadLink);
+    } else {
+      console.warn(
+        "Warning: cannot download the state because state is currently empty."
+      );
+    }
   };
   public autoStepToggle = async () => {
     this.setState({ autostep: !this.state.autostep });
@@ -158,7 +164,7 @@ class App extends React.Component<any, ICanvasState> {
   connectToSocket = () => {
     FileSocket(
       socketAddress,
-      async files => {
+      async (files) => {
         const { domain, substance, style } = files;
         this.setState({ files, connected: true });
 
@@ -230,7 +236,7 @@ class App extends React.Component<any, ICanvasState> {
       history,
       files,
       error,
-      connected
+      connected,
     } = this.state;
     return (
       <div
@@ -239,7 +245,7 @@ class App extends React.Component<any, ICanvasState> {
           height: "100%",
           display: "flex",
           flexFlow: "column",
-          overflow: "hidden"
+          overflow: "hidden",
         }}
       >
         <div style={{ flexShrink: 0 }}>
