+++ packages/core/package.json
@@ -60,6 +60,7 @@
     "@typescript-eslint/parser": "^4.5.0",
     "eslint": "^6.6.0",
     "eslint-config-prettier": "^6.13.0",
+    "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsdoc": "^30.7.3",
     "estrella": "^1.3.0",
     "jest": "^26.6.3",
