+++ packages/core/src/compiler/Substance.test.ts
@@ -5,7 +5,8 @@ import * as path from "path";
 import { PenroseError } from "types/errors";
 import { Result, showError, showType } from "utils/Error";
 import { compileDomain, Env } from "./Domain";
-import { compileSubstance, SubstanceEnv } from "./Substance";
+import { compileSubstance } from "./Substance";
+import { SubstanceEnv } from "types/substance";
 
 const printError = false;
 const saveContexts = false;
