+++ packages/core/src/parser/ParserUtil.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { compact, flatten } from "lodash";
 import * as moo from "moo";
+import { SourceLoc, Identifier } from "types/ast";
 
 export const basicSymbols: moo.Rules = {
   ws: /[ \t]+/,
