+++ packages/core/src/renderer/Renderer.ts
@@ -1,7 +1,9 @@
 import shapeMap from "./shapeMap";
 import { canvasSize } from "renderer/ShapeDef";
-import { Shape } from "types/shapeTypes";
+import { Shape } from "types/shape";
 import { dragUpdate } from "./dragUtils";
+import { IStrV } from "types/value";
+import { LabelCache, State } from "types/state";
 
 export interface ShapeProps {
   shape: Shape;
